requirement.txt
django == 2.1.7
django-mdeditor == 0.1.13
Markdown == 3.0.1
Pygments == 2.3.1
django-suit == 0.2.26